+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This change log follows the conventions of
 
 ## [Unreleased][unreleased]
 
+Nothing so far
+
+## [0.5.2] - 2019-09-02
+
 ### Added
 
 - `CueList` entries can now include nxs2-style DJ comments when they
@@ -503,7 +507,8 @@ This change log follows the conventions of
 - Initial early release of DeviceFinder.
 
 
-[unreleased]: https://github.com/Deep-Symmetry/beat-link/compare/v0.5.1...HEAD
+[unreleased]: https://github.com/Deep-Symmetry/beat-link/compare/v0.5.2...HEAD
+[0.5.2]: https://github.com/Deep-Symmetry/beat-link/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/Deep-Symmetry/beat-link/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/Deep-Symmetry/beat-link/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/Deep-Symmetry/beat-link/compare/v0.4.0...v0.4.1